perf_event: fix warning when compiling with debug enabled
the flags field is an unsigned long, not an int

+++ src/components/perf_event/perf_event.c
@@ -308,7 +308,7 @@ check_exclude_guest( void )
 /*****************************************************************/
 
 static void perf_event_dump_attr( struct perf_event_attr *hw_event,
-	pid_t pid, int cpu, int group_fd, int flags) {
+	pid_t pid, int cpu, int group_fd, unsigned long int flags) {
 
 	/* Mark parameters as not used                   */
 	/* In the common case (no SUBDBG) the function   */
